+++ CHANGELOG.md
@@ -35,6 +35,7 @@
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PoacController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityCoverController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityParticipantController。
 
 - 为 dubbo 的消费者增加了分组设置。
   - snowflake。